backend/requirements.txt
Flask
boto3
flask_cors
python-dotenv